Added http message to Response and Exception classes
Apache on Ubuntu(only confirmed here, 2.2.XX) does not like it if your
header is not in the style:

HTTP/1.1 NNN Short Message

and NNN is not a standard HTTP Code (200, 400, etc). It converts these
all to 500.

Frapi currently does not include the Short Message for custom HTTP
Codes, so I am attempting to add it.